% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{agg.fun}
\alias{agg.fun}
\title{Q Aggregation}
\usage{
agg.fun(B, X.test, y.test, total.step = 10, selection = F)
}
\arguments{
\item{B}{a matrix of estimated coefficients (rows represent predictors, columns represent models)}

\item{X.test}{the test predictor matrix}

\item{y.test}{the test response vector}

\item{total.step}{the maximum number of iterations for the Q-aggregation process (default is 10)}

\item{selection}{a boolean flag; if `TRUE`, selects the coefficients with the smallest prediction error.
If `FALSE`, performs Q-aggregation (default is `FALSE`)}
}
\value{
a list containing:
\item{theta}{the final weight vector for the models}
\item{beta}{the aggregated coefficients}
\item{beta.ew}{the coefficients obtained at the end of the Q-aggregation process (if `selection = FALSE`)}
}
\description{
Aggregation function to estimate coefficients using either model selection or Q-aggregation
}
\examples{
\dontrun{
B <- matrix(rnorm(20), nrow = 4)
X.test <- matrix(rnorm(40), nrow = 10)
y.test <- rnorm(10)
result <- agg.fun(B, X.test, y.test, total.step = 10, selection = FALSE)
print(result)
}
}
