% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betahat}
\alias{betahat}
\title{Compute Beta-Hat Vector}
\usage{
betahat(sigma, u)
}
\arguments{
\item{sigma}{A covariance matrix (Eigen::MatrixXd) used in the computation.}

\item{u}{A numeric vector (Eigen::VectorXd) to be transformed.}
}
\value{
A numeric vector (Eigen::VectorXd) representing the beta-hat, scaled to have a unit norm.
}
\description{
Computes the beta-hat vector using the inverse of the covariance matrix and an input vector.
}
\examples{
\dontrun{
n <- 100
p <- 5
beta_eg <- rep(0.5,p)
Sig.X <- diag(rep(1,p))
X <- rmvnorm(n.vec[k], rep(0, p), Sig.X)
y <- X\%*\%beta_eg + rnorm (n, 0, 1))
u <- un(X,y)
betahat(cov(X),u)
}
}
