% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{fn}
\alias{fn}
\title{FN}
\usage{
fn(beta, beta0)
}
\arguments{
\item{beta}{the estimated coefficients}

\item{beta0}{the true coefficients}
}
\value{
the false negative rate, defined as the proportion of zero entries in `beta` corresponding to nonzero entries in `beta0`
}
\description{
Calculate the false negative rate of an estimator
}
\examples{
\dontrun{
beta <- c(0, 1, 0, 1)
beta0 <- c(1, 1, 0, 1)
fn(beta, beta0)  # Returns 0.3333333
}
}
