% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{fp}
\alias{fp}
\title{FP}
\usage{
fp(beta, beta0)
}
\arguments{
\item{beta}{the estimated coefficients}

\item{beta0}{the true coefficients}
}
\value{
the false positive rate, defined as the proportion of nonzero entries in `beta` corresponding to zero entries in `beta0`
}
\description{
Calculate the false positive rate of an estimator
}
\examples{
\dontrun{
beta <- c(0, 1, 0, 1)
beta0 <- c(0, 0, 0, 1)
fp(beta, beta0)  # Returns 0.5
}
}
