% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{gic_select}
\alias{gic_select}
\title{GIC to select}
\usage{
gic_select(
  lambda,
  u,
  covx,
  n,
  p,
  an,
  start,
  ori,
  weight,
  s = 1,
  L = 1.2,
  alpha = 0.11,
  iter = 200,
  tol = 1e-06
)
}
\arguments{
\item{lambda}{a vector of candidate regularization parameters}

\item{u}{the U-statistics vector}

\item{covx}{the covariance matrix of the predictors}

\item{n}{the number of samples}

\item{p}{the number of predictors}

\item{an}{the penalty adjustment function used in GIC}

\item{start}{the starting point for the optimization}

\item{ori}{the initial estimator using in debias penalty term if theres no need for debias set it 0}

\item{weight}{the weights applied in the lasso optimization}

\item{s}{sign of the first term or u can use sign of U-statistics}

\item{L}{Lipschitz constant used in the optimization algorithm}

\item{alpha}{the learning rate for the optimization}

\item{iter}{the maximum number of iterations for the optimization}

\item{tol}{the tolerance level for convergence}
}
\value{
the best lambda value that minimizes the GIC
}
\description{
Use GIC (Generalized Information Criterion) to select the best regularization parameter
}
\examples{
\dontrun{
gic_select(lambda = c(0.1, 0.2, 0.3), u = c(1, 2),
           covx = matrix(c(1, 0.5, 0.5, 1), 2, 2),
           n = 100, p = 2, an = log, start = c(1, 1),
           ori = c(1, 1), weight = c(1, 1))
}
}
