% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{ind.set}
\alias{ind.set}
\title{index of set}
\usage{
ind.set(n.vec, k.vec)
}
\arguments{
\item{n.vec}{a vector specifying the size of each group; `n.vec[k]` represents the size of the `k`-th group}

\item{k.vec}{a vector of group indices for which the function generates indices}
}
\value{
A vector of indices corresponding to the specified groups
}
\description{
Generate indices for specified groups based on sample splits
}
\examples{
\dontrun{
# Example with 3 groups of sizes 50, 30, and 20
n.vec <- c(50, 30, 20)
# Generate indices for groups 1 and 2
ind.set(n.vec, c(1, 2))
# Output: indices 1 to 80
}
}
