% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{las.kA}
\alias{las.kA}
\title{Oracle Trans-LASSO}
\usage{
las.kA(X, y, A0, n.vec, lam.const = NULL, l1 = T)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{A0}{a set of indices for the initial active set}

\item{n.vec}{a vector defining sample splitting indices; `n.vec[1]` defines the size of the first part}

\item{lam.const}{an optional constant for the penalty term; if `NULL`, it is determined through cross-validation}

\item{l1}{a logical flag indicating whether to use lasso regression in the first stage (default is `TRUE`)}
}
\value{
a list containing:
\item{beta.kA}{the estimated coefficients after the two-stage procedure}
\item{w.kA}{the coefficients from the first stage of lasso regression (if applicable)}
\item{lam.const}{the constant for the penalty term}
}
\description{
Oracle Trans-LASSO
}
\examples{
\dontrun{
X <- matrix(rnorm(100 * 10), 100, 10)
y <- rnorm(100)
A0 <- c(1, 2, 3)
n.vec <- c(80, 20)
result <- las.kA(X, y, A0, n.vec)
print(result)
}
}
