% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{mse.fun}
\alias{mse.fun}
\title{mse}
\usage{
mse.fun(beta, est, X.test = NULL)
}
\arguments{
\item{beta}{the true coefficients}

\item{est}{the estimated coefficients}

\item{X.test}{(optional) the test design matrix}
}
\value{
a list containing:
\item{est.err}{the estimation error (squared error between `beta` and `est`)}
\item{pred.err}{the prediction error (mean squared error of predictions), or `NA` if `X.test` is not provided}
}
\description{
Compute estimation error and prediction error (optional)
}
\examples{
\dontrun{
beta <- c(1, 0, 0.5)
est <- c(1.1, 0, 0.4)
X.test <- matrix(c(1, 0.5, -0.2, 1.5, 0.3, 0.7), 3, 2)
mse.fun(beta, est, X.test)
}
}
