% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{relu}
\alias{relu}
\title{ReLU Activation Function}
\usage{
relu(vec)
}
\arguments{
\item{vec}{A numeric vector (Eigen::VectorXd) for which the ReLU activation is computed.}
}
\value{
A numeric vector (Eigen::VectorXd) with the same dimensions as `vec`, where each element is replaced by `max(0, element)`.
}
\description{
Computes the Rectified Linear Unit (ReLU) activation for a given numeric vector.
}
\examples{
\dontrun{
# Example vector
input_vec <- c(-2, -1, 0, 1, 2)
# Compute ReLU
output_vec <- relu(input_vec)
print(output_vec)  # Should return c(0, 0, 0, 1, 2)
}
}
