% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{repcol}
\alias{repcol}
\title{rep columns}
\usage{
repcol(x, n)
}
\arguments{
\item{x}{a numeric vector to be replicated}

\item{n}{the number of columns to replicate}
}
\value{
A matrix with `n` columns, each containing the replicated values of `x`
}
\description{
Replicate a vector as columns in a matrix
}
\examples{
\dontrun{
# Replicate the vector c(1, 2, 3) into 4 columns
repcol(c(1, 2, 3), 4)
# Output:
#      [,1] [,2] [,3] [,4]
# [1,]    1    1    1    1
# [2,]    2    2    2    2
# [3,]    3    3    3    3
}
}
