% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{Trans.lasso.sp}
\alias{Trans.lasso.sp}
\title{Trans.lasso.sp}
\usage{
Trans.lasso.sp(X, y, n.vec, I.til, l1 = T)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{n.vec}{a vector defining the sample splitting indices; `n.vec[k]` defines the size of the `k`th split}

\item{I.til}{a vector of indices for the transfer dataset to be used for aggregation}

\item{l1}{a boolean flag indicating whether to use lasso regression in the selection steps (default is `TRUE`)}
}
\value{
a list containing:
\item{beta.sp}{the selected coefficients using model selection aggregation}
\item{theta.sp}{the weights assigned to models during aggregation}
\item{rank.pi}{the ranks of transfer datasets based on the computed Rhat values}
}
\description{
Perform Transfer Learning with a sparsity constraint using Lasso regression
}
\examples{
\dontrun{
X <- matrix(rnorm(1000), 100, 10)
y <- rnorm(100)
n.vec <- c(50, 30, 20)
I.til <- sample(1:50, 10)
result <- Trans.lasso.sp(X, y, n.vec, I.til)
print(result)
}
}
