\name{TransSIM-package}
\alias{TransSIM-package}
\alias{TransSIM}
\docType{package}
\title{
  Transfer learning for Single-Index Models
}
\description{
  Transfer learning for single-index models.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Yanlong Zhao, zhaoyanlong@mail.ustc.edu.cn.

Maintainer: Yanlong Zhao <zhaoyanlong@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
